requirements.txt
streamlit
pandas
gspread
google-auth
openpyxl
xlsxwriter
xlsxwriter
pyxlsb
tabula-py  # opcional, só se tiver Java no ambiente
PyPDF2
